requirements.txt
annotated-types==0.6.0
anyio==4.2.0
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
ghp-import==2.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-material==9.5.8
mkdocs-material-extensions==1.3.1
mkdocs-static-i18n==1.2.0
openai==1.11.1
packaging==23.2
paginate==0.5.6
pathspec==0.12.1
platformdirs==4.2.0
pyaml==23.12.0
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pymdown-extensions==10.7
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
watchdog==4.0.0
